Basic round functionality created and working.
Need to be able to view rounds but users can now start a round and add orders. People or drinks not already known to the system are rejected.
Want to be able to add a new person/drink to all known people/dirnks instead of rejection.
Persisting rounds in memory not yet implemented.